refactor(ResultPage): move vorabcheck results direct to ResultPage component
