--- app/components/resultPage/FluggastrechteResultPage.tsx
@@ -1,17 +1,0 @@
-import { useLoaderData } from "@remix-run/react";
-import { ResultPage } from "./ResultPage";
-import type { loader } from "../../routes/shared/result.server";
-
-export function FluggastrechteResultPage() {
-  const { common, cmsData, reasons, backButton } =
-    useLoaderData<typeof loader>();
-
-  return (
-    <ResultPage
-      common={common}
-      cmsData={cmsData}
-      reasons={reasons}
-      backButton={backButton}
-    />
-  );
-}
