+++ app/routes/beratungshilfe.vorabcheck.ergebnis.$.tsx
@@ -1,2 +1,2 @@
 export { loader } from "~/routes/shared/result.server";
-export { DefaultResultPage as default } from "../components/resultPage/DefaultResultPage";
+export { ResultPage as default } from "../components/resultPage/ResultPage";
