+++ app/routes/geld-einklagen.vorabcheck.ergebnis.$.tsx
@@ -2,7 +2,7 @@ import type { LoaderFunctionArgs } from "@remix-run/node";
 import { loader as stepLoader } from "~/routes/shared/result.server";
 import { throw404IfFeatureFlagDisabled } from "~/services/errorPages/throw404";
 
-export { DefaultResultPage as default } from "../components/resultPage/DefaultResultPage";
+export { ResultPage as default } from "../components/resultPage/ResultPage";
 
 export const loader = async (opts: LoaderFunctionArgs) => {
   await throw404IfFeatureFlagDisabled("showGeldEinklagenFlow");
